# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      message:
        example: Example
        type: string
    type: object
  controllers.FileResponse:
    properties:
      id:
        type: integer
      url:
        type: string
    type: object
...
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API для BlogPoint
  title: BlogPoint API
  version: "1.0"
paths:
...
  /api/deleteFile/{id}:
    delete:
      description: Удаляет файл по id
      parameters:
      - description: Id файла для удаления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление файла
      tags:
      - File
...
  /api/uploadFile:
    post:
      consumes:
      - multipart/form-data
      description: Загружает файл и возвращает его URL и уникальное имя
      parameters:
      - description: Файл для загрузки
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.FileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузка файла
      tags:
      - File
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: jwt
    type: apiKey
swagger: "2.0"